removed redundant test script
